announcements api and services file split, tested and works well
